refactor modules html to templates fix tests
