+++ lib/oli_web/controllers/lti_controller.ex
@@ -2,8 +2,6 @@ defmodule OliWeb.LtiController do
   use OliWeb, :controller
   use OliWeb, :verified_routes
 
-  import Oli.Utils
-
   alias Oli.Accounts
   alias Oli.Delivery.Sections
   alias Oli.Institutions
@@ -17,6 +15,8 @@ defmodule OliWeb.LtiController do
   alias Oli.Delivery.Attempts.Core
   alias Oli.Delivery.Attempts.Core.ResourceAttempt
   alias Oli.Lti.PlatformInstances
+  alias Oli.Lti.Tokens
+  alias Oli.Lti.PlatformExternalTools
   alias Lti_1p3.Roles.ContextRoles
   alias Lti_1p3.Roles.PlatformRoles
   alias Lti_1p3.Tool.Services.AGS
@@ -79,6 +79,116 @@ defmodule OliWeb.LtiController do
     end
   end
 
+  defp handle_valid_lti_1p3_launch(conn, lti_params) do
+    issuer = lti_params["iss"]
+    client_id = LtiParams.peek_client_id(lti_params)
+    deployment_id = lti_params["https://purl.imsglobal.org/spec/lti/claim/deployment_id"]
+
+    Oli.Repo.transaction(fn ->
+      case Institutions.get_institution_registration_deployment(issuer, client_id, deployment_id) do
+        {institution, registration, _deployment} ->
+          lti_roles = lti_params["https://purl.imsglobal.org/spec/lti/claim/roles"]
+
+          # update user values defined by the oidc standard per LTI 1.3 standard user identity claims
+          # http://www.imsglobal.org/spec/lti/v1p3/#user-identity-claims
+          case Accounts.insert_or_update_lms_user(
+                 %{
+                   sub: lti_params["sub"],
+                   name: lti_params["name"],
+                   given_name: lti_params["given_name"],
+                   family_name: lti_params["family_name"],
+                   middle_name: lti_params["middle_name"],
+                   nickname: lti_params["nickname"],
+                   preferred_username: lti_params["preferred_username"],
+                   profile: lti_params["profile"],
+                   picture: lti_params["picture"],
+                   website: lti_params["website"],
+                   email: lti_params["email"],
+                   email_verified: true,
+                   gender: lti_params["gender"],
+                   birthdate: lti_params["birthdate"],
+                   zoneinfo: lti_params["zoneinfo"],
+                   locale: lti_params["locale"],
+                   phone_number: lti_params["phone_number"],
+                   phone_number_verified: lti_params["phone_number_verified"],
+                   address: lti_params["address"],
+                   lti_institution_id: institution.id
+                 },
+                 institution.id
+               ) do
+            {:ok, user} ->
+              # update lti params and session to be associated with the current lms user
+              {:ok, _} =
+                LtiParams.create_or_update_lti_params(lti_params, user.id)
+
+              # update user platform roles
+              Accounts.update_user_platform_roles(
+                user,
+                PlatformRoles.get_roles_by_uris(lti_roles)
+              )
+
+              # context claim is considered optional according to IMS http://www.imsglobal.org/spec/lti/v1p3/#context-claim
+              # safeguard against the case that context is missing
+              case lti_params["https://purl.imsglobal.org/spec/lti/claim/context"] do
+                nil ->
+                  {_error_id, error_msg} =
+                    Oli.Utils.log_error("context claim is missing from lti params")
+
+                  throw(error_msg)
+
+                context ->
+                  # update section specifics - if one exists. Enroll the user and also update the section details
+                  with {:ok, section} <- get_existing_section(lti_params) do
+                    # transform lti_roles to a list only containing valid context roles (exclude all system and institution roles)
+                    context_roles = ContextRoles.get_roles_by_uris(lti_roles)
+
+                    # if a course section exists, ensure that this user has an enrollment in this section
+                    enroll_user(user.id, section.id, context_roles)
+
+                    # make sure section details are up to date
+                    %{"title" => context_title} = context
+
+                    {:ok, _section} =
+                      update_section_details(context_title, section, lti_params, registration)
+                  end
+
+                  # sign current user in and redirect to home page
+                  conn
+                  |> UserAuth.create_session(user)
+                  |> redirect(to: "/sections")
+              end
+
+            {:error, changeset} ->
+              {_error_id, error_msg} =
+                Oli.Utils.log_error("Failed to create or update user", changeset)
+
+              throw(error_msg)
+          end
+      end
+    end)
+  end
+
+  defp enroll_user(user_id, section_id, context_roles) do
+    Sections.enroll(user_id, section_id, context_roles)
+  end
+
+  defp update_section_details(context_title, section, lti_params, registration) do
+    Sections.update_section(section, %{
+      title: context_title,
+      grade_passback_enabled: AGS.grade_passback_enabled?(lti_params),
+      line_items_service_url: AGS.get_line_items_url(lti_params, registration),
+      nrps_enabled: NRPS.nrps_enabled?(lti_params),
+      nrps_context_memberships_url: NRPS.get_context_memberships_url(lti_params)
+    })
+  end
+
+  defp get_existing_section(lti_params) do
+    case Sections.get_section_from_lti_params(lti_params) do
+      nil -> nil
+      section -> {:ok, section}
+    end
+  end
+
   def test(conn, params) do
     session_state = Plug.Conn.get_session(conn, "state")
 
@@ -101,7 +211,7 @@ defmodule OliWeb.LtiController do
       %Lti_1p3.Platform.LoginHint{context: context, session_user_id: _session_user_id} ->
         client_id = params["client_id"]
 
-        platform_instance =
+        {:ok, platform_instance} =
           PlatformInstances.get_platform_instance_by_client_id(client_id)
 
         {user, target_link_uri, activity_resource_id, roles, additional_claims, message_type} =
@@ -393,6 +503,137 @@ defmodule OliWeb.LtiController do
       Lti_1p3.Claims.Custom.custom(custom)
     ]
 
+  @doc """
+  Handles the Deep Linking response from the LTI tool.
+  This endpoint receives the JWT containing the content items selected by the user.
+  """
+  def deep_link(
+        conn,
+        %{"JWT" => jwt, "section_slug" => section_slug, "resource_id" => resource_id} = _params
+      ) do
+    with {:ok, claims} <- validate_deep_linking_jwt(jwt),
+         {:ok, content_item} <- require_single_content_item(claims),
+         :ok <- require_lti_resource_link_type(content_item),
+         {:ok, section} <- get_section_by_slug(section_slug),
+         :ok = process_deep_linking_content_item(content_item, section.id, resource_id) do
+      # Render success page with postMessage communication
+      conn
+      |> put_view(OliWeb.LtiHTML)
+      |> put_layout(false)
+      |> put_format("html")
+      |> render(:deep_link_success, content_item: content_item)
+    else
+      {:error, _reason, error_description} ->
+        conn
+        |> put_view(OliWeb.LtiHTML)
+        |> put_layout(false)
+        |> put_format("html")
+        |> put_status(:bad_request)
+        |> render(:deep_link_error, error_description: error_description)
+
+      {:error, reason} ->
+        error_description = to_string(reason)
+
+        conn
+        |> put_view(OliWeb.LtiHTML)
+        |> put_layout(false)
+        |> put_format("html")
+        |> put_status(:bad_request)
+        |> render(:deep_link_error, error_description: error_description)
+    end
+  end
+
+  defp get_section_by_slug(section_slug) do
+    case Sections.get_section_by_slug(section_slug) do
+      %Sections.Section{} = section ->
+        {:ok, section}
+
+      nil ->
+        {:error, :section_not_found}
+    end
+  end
+
+  defp validate_deep_linking_jwt(jwt) do
+    with {:ok, %JOSE.JWT{fields: %{"iss" => client_id, "aud" => audience}}} <-
+           Tokens.peek_jwt(jwt),
+         %JOSE.JWS{fields: %{"kid" => kid}} <- JOSE.JWT.peek_protected(jwt),
+         {:ok, platform_instance} <-
+           PlatformInstances.get_platform_instance_by_client_id(client_id),
+         {:ok, jwk} <- Tokens.get_jwk_for_assertion(platform_instance.keyset_url, kid),
+         {true, jwt, _jws} <- JOSE.JWT.verify(jwk, jwt),
+         jwt_claims <- jwt.fields,
+         :ok <- validate_audience(audience),
+         :ok <- validate_message_type(jwt_claims) do
+      {:ok, jwt_claims}
+    else
+      e ->
+        Logger.error("Failed to validate deep linking JWT: #{inspect(e)}")
+        {:error, :invalid_deep_linking_jwt}
+    end
+  end
+
+  defp validate_audience(audience) when is_list(audience),
+    do: audience |> hd() |> validate_audience()
+
+  defp validate_audience(audience) do
+    expected_audience = Oli.Utils.get_base_url()
+
+    if audience == expected_audience do
+      :ok
+    else
+      {:error, :invalid_audience}
+    end
+  end
+
+  defp validate_message_type(%{
+         "https://purl.imsglobal.org/spec/lti/claim/message_type" => "LtiDeepLinkingResponse"
+       }) do
+    :ok
+  end
+
+  defp validate_message_type(_) do
+    {:error, :invalid_message_type}
+  end
+
+  defp require_single_content_item(%{
+         "https://purl.imsglobal.org/spec/lti-dl/claim/content_items" => content_items
+       }) do
+    if is_list(content_items) and length(content_items) == 1 do
+      {:ok, hd(content_items)}
+    else
+      {:error, :invalid_content_item,
+       "Expected exactly one content item, got #{length(content_items)}"}
+    end
+  end
+
+  defp require_lti_resource_link_type(%{"type" => "ltiResourceLink"}), do: :ok
+
+  defp require_lti_resource_link_type(_),
+    do: {:error, :invalid_content_item_type, "Expected content item type to be 'ltiResourceLink'"}
+
+  defp process_deep_linking_content_item(content_item, section_id, resource_id) do
+    Logger.info("Processing deep linking content item: #{inspect(content_item)}")
+
+    case PlatformExternalTools.upsert_section_resource_deep_link(%{
+           type: content_item["type"],
+           title: content_item["title"],
+           text: content_item["text"],
+           url: content_item["url"],
+           custom: content_item["custom"],
+           resource_id: resource_id,
+           section_id: section_id
+         }) do
+      {:ok, _deep_link} ->
+        :ok
+
+      {:error, reason} ->
+        Logger.error("Failed to create section resource deep link: #{inspect(reason)}")
+        {:error, :failed_to_create_deep_link}
+    end
+
+    :ok
+  end
+
   def request_registration(
         conn,
         %{"pending_registration" => pending_registration_attrs} = _params
@@ -502,114 +743,6 @@ defmodule OliWeb.LtiController do
       pending_registration ->
         conn
         |> render("registration_pending.html", pending_registration: pending_registration)
-    end
-  end
-
-  defp handle_valid_lti_1p3_launch(conn, lti_params) do
-    issuer = lti_params["iss"]
-    client_id = LtiParams.peek_client_id(lti_params)
-    deployment_id = lti_params["https://purl.imsglobal.org/spec/lti/claim/deployment_id"]
-
-    Oli.Repo.transaction(fn ->
-      case Institutions.get_institution_registration_deployment(issuer, client_id, deployment_id) do
-        {institution, registration, _deployment} ->
-          lti_roles = lti_params["https://purl.imsglobal.org/spec/lti/claim/roles"]
-
-          # update user values defined by the oidc standard per LTI 1.3 standard user identity claims
-          # http://www.imsglobal.org/spec/lti/v1p3/#user-identity-claims
-          case Accounts.insert_or_update_lms_user(
-                 %{
-                   sub: lti_params["sub"],
-                   name: lti_params["name"],
-                   given_name: lti_params["given_name"],
-                   family_name: lti_params["family_name"],
-                   middle_name: lti_params["middle_name"],
-                   nickname: lti_params["nickname"],
-                   preferred_username: lti_params["preferred_username"],
-                   profile: lti_params["profile"],
-                   picture: lti_params["picture"],
-                   website: lti_params["website"],
-                   email: lti_params["email"],
-                   email_verified: true,
-                   gender: lti_params["gender"],
-                   birthdate: lti_params["birthdate"],
-                   zoneinfo: lti_params["zoneinfo"],
-                   locale: lti_params["locale"],
-                   phone_number: lti_params["phone_number"],
-                   phone_number_verified: lti_params["phone_number_verified"],
-                   address: lti_params["address"],
-                   lti_institution_id: institution.id
-                 },
-                 institution.id
-               ) do
-            {:ok, user} ->
-              # update lti params and session to be associated with the current lms user
-              {:ok, _} =
-                LtiParams.create_or_update_lti_params(lti_params, user.id)
-
-              # update user platform roles
-              Accounts.update_user_platform_roles(
-                user,
-                PlatformRoles.get_roles_by_uris(lti_roles)
-              )
-
-              # context claim is considered optional according to IMS http://www.imsglobal.org/spec/lti/v1p3/#context-claim
-              # safeguard against the case that context is missing
-              case lti_params["https://purl.imsglobal.org/spec/lti/claim/context"] do
-                nil ->
-                  {_error_id, error_msg} = log_error("context claim is missing from lti params")
-
-                  throw(error_msg)
-
-                context ->
-                  # update section specifics - if one exists. Enroll the user and also update the section details
-                  with {:ok, section} <- get_existing_section(lti_params) do
-                    # transform lti_roles to a list only containing valid context roles (exclude all system and institution roles)
-                    context_roles = ContextRoles.get_roles_by_uris(lti_roles)
-
-                    # if a course section exists, ensure that this user has an enrollment in this section
-                    enroll_user(user.id, section.id, context_roles)
-
-                    # make sure section details are up to date
-                    %{"title" => context_title} = context
-
-                    {:ok, _section} =
-                      update_section_details(context_title, section, lti_params, registration)
-                  end
-
-                  # sign current user in and redirect to home page
-                  conn
-                  |> UserAuth.create_session(user)
-                  |> redirect(to: "/sections")
-              end
-
-            {:error, changeset} ->
-              {_error_id, error_msg} = log_error("Failed to create or update user", changeset)
-
-              throw(error_msg)
-          end
-      end
-    end)
-  end
-
-  defp enroll_user(user_id, section_id, context_roles) do
-    Sections.enroll(user_id, section_id, context_roles)
-  end
-
-  defp update_section_details(context_title, section, lti_params, registration) do
-    Sections.update_section(section, %{
-      title: context_title,
-      grade_passback_enabled: AGS.grade_passback_enabled?(lti_params),
-      line_items_service_url: AGS.get_line_items_url(lti_params, registration),
-      nrps_enabled: NRPS.nrps_enabled?(lti_params),
-      nrps_context_memberships_url: NRPS.get_context_memberships_url(lti_params)
-    })
-  end
-
-  defp get_existing_section(lti_params) do
-    case Sections.get_section_from_lti_params(lti_params) do
-      nil -> nil
-      section -> {:ok, section}
     end
   end
 end
