+++ lib/oli_web/controllers/lti_html/deep_link_error.html.heex
@@ -1,0 +1,33 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Deep Linking Error</title>
+    <style>
+      body {
+        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+        padding: 20px;
+        text-align: center;
+      }
+      .error {
+        color: #ef4444;
+        margin: 20px 0;
+      }
+    </style>
+  </head>
+  <body>
+    <div class="error">
+      <h2>Deep Linking Failed</h2>
+      <p><%= @error_description %></p>
+    </div>
+    <script>
+      if (window.parent !== window) {
+        window.parent.postMessage({
+          type: 'lti_deep_linking_response',
+          status: 'error',
+          error: 'invalid_request',
+          error_description: '<%= @error_description %>'
+        }, '*');
+      }
+    </script>
+  </body>
+</html>
