+++ lib/oli_web/controllers/lti_html/deep_link_success.html.heex
@@ -1,0 +1,55 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Deep Linking Complete</title>
+    <style>
+      body {
+        font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+        padding: 20px;
+        text-align: center;
+      }
+      .success {
+        color: #22c55e;
+        margin: 20px 0;
+      }
+      .content-item {
+        background: #f8fafc;
+        padding: 15px;
+        border-radius: 8px;
+        margin: 10px 0;
+      }
+    </style>
+  </head>
+  <body>
+    <div class="success">
+      <h2>Resource Selected Successfully</h2>
+      <div class="content-item">
+        <h3><%= @content_item["title"] || "Selected Resource" %></h3>
+        <%= if @content_item["text"] do %>
+          <p><%= @content_item["text"] %></p>
+        <% end %>
+        <%= if @content_item["url"] do %>
+          <p><strong>URL:</strong> <%= @content_item["url"] %></p>
+        <% end %>
+      </div>
+      <p>This window will close automatically...</p>
+    </div>
+    <script>
+      if (window.parent !== window) {
+        window.parent.postMessage({
+          type: 'lti_deep_linking_response',
+          status: 'success',
+          content_item: <%= raw Jason.encode!(@content_item) %>
+        }, '*');
+
+        window.parent.postMessage({
+          type: 'lti_close_modal'
+        }, '*');
+
+        setTimeout(() => {
+          window.close();
+        }, 2000);
+      }
+    </script>
+  </body>
+</html>
