+++ test/oli_web/controllers/lti_controller_test.exs
@@ -7,6 +7,10 @@ defmodule OliWeb.LtiControllerTest do
   alias Oli.Delivery.Sections
   alias Oli.Accounts.User
   alias Oli.Lti.PlatformExternalTools
+  alias Oli.Authoring.Editing.{ActivityEditor, PageEditor}
+  alias Oli.Publishing
+  alias Oli.Delivery.Sections
+  alias Oli.Test.MockHTTP
 
   import Mox
   import Oli.Factory
@@ -487,102 +491,442 @@ defmodule OliWeb.LtiControllerTest do
       assert html_response(conn, 200) =~
                "<form name=\"post_redirect\" action=\"#{target_link_uri}\" method=\"post\">"
     end
+  end
 
-    test "returns developer key json", %{conn: conn} do
-      conn = get(conn, Routes.lti_path(conn, :developer_key_json))
+  describe "deep_link" do
+    setup [:setup_section]
 
-      {:ok, active_jwk} = Lti_1p3.get_active_jwk()
+    defp assert_deep_link_error_response(resp, expected_error_text) do
+      # Verify HTML error response contains error message
+      assert resp =~ "Deep Linking Failed"
 
-      public_jwk =
-        JOSE.JWK.from_pem(active_jwk.pem)
-        |> JOSE.JWK.to_public()
-        |> JOSE.JWK.to_map()
-        |> (fn {_kty, public_jwk} -> public_jwk end).()
-        |> Map.put("kid", active_jwk.kid)
+      if expected_error_text do
+        assert resp =~ expected_error_text
+      end
 
-      assert json_response(conn, 200) != nil
-
-      assert json_response(conn, 200)
-             |> Map.get("extensions")
-             |> Enum.find(fn %{"platform" => p} -> p == "canvas.instructure.com" end)
-             |> Map.get("privacy_level") =~ "public"
-
-      assert json_response(conn, 200) |> Map.get("title") =~ "OLI Torus"
-
-      assert json_response(conn, 200) |> Map.get("description") =~
-               "Create, deliver and iteratively improve course content"
-
-      assert json_response(conn, 200) |> Map.get("oidc_initiation_url") =~
-               "https://localhost/lti/login"
-
-      assert json_response(conn, 200) |> Map.get("target_link_uri") =~
-               "https://localhost/lti/launch"
-
-      assert json_response(conn, 200) |> Map.get("public_jwk_url") =~
-               "https://localhost/.well-known/jwks.json"
-
-      assert json_response(conn, 200) |> Map.get("scopes") == [
-               "https://purl.imsglobal.org/spec/lti-ags/scope/lineitem",
-               "https://purl.imsglobal.org/spec/lti-ags/scope/lineitem.readonly",
-               "https://purl.imsglobal.org/spec/lti-ags/scope/result.readonly",
-               "https://purl.imsglobal.org/spec/lti-ags/scope/score",
-               "https://purl.imsglobal.org/spec/lti-nrps/scope/contextmembership.readonly"
-             ]
-
-      %{"extensions" => [%{"settings" => %{"placements" => placements}} | _]} =
-        json_response(conn, 200)
-
-      assert placements == [
-               %{
-                 "icon_url" => "https://localhost/branding/prod/oli_torus_icon.png",
-                 "message_type" => "LtiResourceLinkRequest",
-                 "placement" => "link_selection"
-               },
-               %{
-                 "message_type" => "LtiResourceLinkRequest",
-                 "placement" => "assignment_selection"
-               },
-               %{
-                 "message_type" => "LtiResourceLinkRequest",
-                 "placement" => "course_navigation",
-                 "default" => "disabled",
-                 "windowTarget" => "_blank"
-               }
-             ]
-
-      assert json_response(conn, 200) |> Map.get("public_jwk") |> Map.get("kid") ==
-               public_jwk["kid"]
-
-      assert json_response(conn, 200) |> Map.get("public_jwk") |> Map.get("n") == public_jwk["n"]
+      # Verify error postMessage JavaScript is included
+      assert resp =~ "window.parent.postMessage"
+      assert resp =~ "lti_deep_linking_response"
+      assert resp =~ "status: 'error'"
     end
 
-    test "returns developer key json with course navigation enabled", %{conn: conn} do
-      conn =
-        get(
-          conn,
-          Routes.lti_path(conn, :developer_key_json, course_navigation_default: "enabled")
+    defp generate_deep_linking_jwt(client_id, key, kid, claims \\ %{}) do
+      now = DateTime.utc_now() |> DateTime.to_unix()
+
+      base_claims = %{
+        "iss" => client_id,
+        "aud" => Oli.Utils.get_base_url(),
+        "iat" => now,
+        "exp" => now + 3600,
+        "jti" => UUID.uuid4(),
+        "https://purl.imsglobal.org/spec/lti/claim/message_type" => "LtiDeepLinkingResponse",
+        "https://purl.imsglobal.org/spec/lti-dl/claim/content_items" => [
+          %{
+            "type" => "ltiResourceLink",
+            "title" => "Test Resource",
+            "text" => "A test resource for deep linking",
+            "url" => "https://example.com/resource/123",
+            "custom" => %{
+              "param1" => "value1",
+              "param2" => "value2"
+            }
+          }
+        ]
+      }
+
+      final_claims = Map.merge(base_claims, claims)
+
+      jwk = JOSE.JWK.from_pem(key)
+
+      {_, jwt} =
+        JOSE.JWT.sign(
+          jwk,
+          %{"alg" => "RS256", "kid" => kid},
+          final_claims
         )
+        |> JOSE.JWS.compact()
 
-      %{"extensions" => [%{"settings" => %{"placements" => placements}} | _]} =
-        json_response(conn, 200)
+      jwt
+    end
 
-      assert placements == [
-               %{
-                 "icon_url" => "https://localhost/branding/prod/oli_torus_icon.png",
-                 "message_type" => "LtiResourceLinkRequest",
-                 "placement" => "link_selection"
-               },
-               %{
-                 "message_type" => "LtiResourceLinkRequest",
-                 "placement" => "assignment_selection"
-               },
-               %{
-                 "message_type" => "LtiResourceLinkRequest",
-                 "placement" => "course_navigation",
-                 "default" => "enabled",
-                 "windowTarget" => "_blank"
-               }
-             ]
+    test "successfully processes valid deep linking response", %{
+      conn: conn,
+      section: section,
+      activity_id: activity_id
+    } do
+      client_id = "test-client-id"
+      key_pem = File.read!("test/support/fixtures/test_rsa_private.pem")
+
+      public_jwk =
+        JOSE.JWK.from_pem(key_pem)
+        |> JOSE.JWK.to_public()
+        |> JOSE.JWK.to_map()
+        |> elem(1)
+        |> Map.put("kid", "test-kid")
+
+      keyset_url = "https://some-tool.example.edu/.well-known/jwks.json"
+
+      # Insert a platform instance
+      insert(:platform_instance, %{
+        client_id: client_id,
+        keyset_url: keyset_url
+      })
+
+      # Mock HTTP request to fetch the keyset
+      MockHTTP
+      |> expect(:get, fn ^keyset_url ->
+        {:ok,
+         %HTTPoison.Response{
+           status_code: 200,
+           body: Jason.encode!(%{"keys" => [public_jwk]})
+         }}
+      end)
+
+      jwt = generate_deep_linking_jwt(client_id, key_pem, "test-kid")
+
+      params = %{
+        "JWT" => jwt
+      }
+
+      conn = post(conn, ~p"/lti/deep_link/#{section.slug}/#{activity_id}", params)
+      resp = html_response(conn, 200)
+
+      # Verify HTML response contains success message
+      assert resp =~ "Resource Selected Successfully"
+      assert resp =~ "Test Resource"
+      # Verify postMessage JavaScript is included
+      assert resp =~ "window.parent.postMessage"
+      assert resp =~ "lti_deep_linking_response"
+      assert resp =~ "lti_close_modal"
+    end
+
+    @tag capture_log: true
+    test "returns 400 for invalid JWT", %{conn: conn, section: section, activity_id: activity_id} do
+      params = %{
+        "JWT" => "invalid.jwt.token"
+      }
+
+      conn = post(conn, ~p"/lti/deep_link/#{section.slug}/#{activity_id}", params)
+      resp = html_response(conn, 400)
+
+      assert_deep_link_error_response(resp, "invalid_deep_linking_jwt")
+    end
+
+    @tag capture_log: true
+    test "returns 400 for JWT with wrong message type", %{
+      conn: conn,
+      section: section,
+      activity_id: activity_id
+    } do
+      client_id = "test-client-id"
+      key_pem = File.read!("test/support/fixtures/test_rsa_private.pem")
+
+      public_jwk =
+        JOSE.JWK.from_pem(key_pem)
+        |> JOSE.JWK.to_public()
+        |> JOSE.JWK.to_map()
+        |> elem(1)
+        |> Map.put("kid", "test-kid")
+
+      keyset_url = "https://some-tool.example.edu/.well-known/jwks.json"
+
+      insert(:platform_instance, %{
+        client_id: client_id,
+        keyset_url: keyset_url
+      })
+
+      MockHTTP
+      |> expect(:get, fn ^keyset_url ->
+        {:ok,
+         %HTTPoison.Response{
+           status_code: 200,
+           body: Jason.encode!(%{"keys" => [public_jwk]})
+         }}
+      end)
+
+      # Generate JWT with wrong message type
+      jwt =
+        generate_deep_linking_jwt(client_id, key_pem, "test-kid", %{
+          "https://purl.imsglobal.org/spec/lti/claim/message_type" => "LtiResourceLinkRequest"
+        })
+
+      params = %{
+        "JWT" => jwt
+      }
+
+      conn = post(conn, ~p"/lti/deep_link/#{section.slug}/#{activity_id}", params)
+      resp = html_response(conn, 400)
+
+      assert_deep_link_error_response(resp, "invalid_deep_linking_jwt")
+    end
+
+    @tag capture_log: true
+    test "returns 400 for JWT with wrong audience", %{
+      conn: conn,
+      section: section,
+      activity_id: activity_id
+    } do
+      client_id = "test-client-id"
+      key_pem = File.read!("test/support/fixtures/test_rsa_private.pem")
+
+      public_jwk =
+        JOSE.JWK.from_pem(key_pem)
+        |> JOSE.JWK.to_public()
+        |> JOSE.JWK.to_map()
+        |> elem(1)
+        |> Map.put("kid", "test-kid")
+
+      keyset_url = "https://some-tool.example.edu/.well-known/jwks.json"
+
+      insert(:platform_instance, %{
+        client_id: client_id,
+        keyset_url: keyset_url
+      })
+
+      MockHTTP
+      |> expect(:get, fn ^keyset_url ->
+        {:ok,
+         %HTTPoison.Response{
+           status_code: 200,
+           body: Jason.encode!(%{"keys" => [public_jwk]})
+         }}
+      end)
+
+      # Generate JWT with wrong audience
+      jwt =
+        generate_deep_linking_jwt(client_id, key_pem, "test-kid", %{
+          "aud" => "https://wrong-audience.com"
+        })
+
+      params = %{
+        "JWT" => jwt
+      }
+
+      conn = post(conn, ~p"/lti/deep_link/#{section.slug}/#{activity_id}", params)
+      resp = html_response(conn, 400)
+
+      assert_deep_link_error_response(resp, "invalid_deep_linking_jwt")
+    end
+
+    test "returns 400 for JWT with multiple content items", %{
+      conn: conn,
+      section: section,
+      activity_id: activity_id
+    } do
+      client_id = "test-client-id"
+      key_pem = File.read!("test/support/fixtures/test_rsa_private.pem")
+
+      public_jwk =
+        JOSE.JWK.from_pem(key_pem)
+        |> JOSE.JWK.to_public()
+        |> JOSE.JWK.to_map()
+        |> elem(1)
+        |> Map.put("kid", "test-kid")
+
+      keyset_url = "https://some-tool.example.edu/.well-known/jwks.json"
+
+      insert(:platform_instance, %{
+        client_id: client_id,
+        keyset_url: keyset_url
+      })
+
+      MockHTTP
+      |> expect(:get, fn ^keyset_url ->
+        {:ok,
+         %HTTPoison.Response{
+           status_code: 200,
+           body: Jason.encode!(%{"keys" => [public_jwk]})
+         }}
+      end)
+
+      # Generate JWT with multiple content items
+      jwt =
+        generate_deep_linking_jwt(client_id, key_pem, "test-kid", %{
+          "https://purl.imsglobal.org/spec/lti-dl/claim/content_items" => [
+            %{
+              "type" => "ltiResourceLink",
+              "title" => "Test Resource 1",
+              "url" => "https://example.com/resource/1"
+            },
+            %{
+              "type" => "ltiResourceLink",
+              "title" => "Test Resource 2",
+              "url" => "https://example.com/resource/2"
+            }
+          ]
+        })
+
+      params = %{
+        "JWT" => jwt
+      }
+
+      conn = post(conn, ~p"/lti/deep_link/#{section.slug}/#{activity_id}", params)
+      resp = html_response(conn, 400)
+
+      assert_deep_link_error_response(resp, "Expected exactly one content item, got 2")
+    end
+
+    test "returns 400 for JWT with no content items", %{
+      conn: conn,
+      section: section,
+      activity_id: activity_id
+    } do
+      client_id = "test-client-id"
+      key_pem = File.read!("test/support/fixtures/test_rsa_private.pem")
+
+      public_jwk =
+        JOSE.JWK.from_pem(key_pem)
+        |> JOSE.JWK.to_public()
+        |> JOSE.JWK.to_map()
+        |> elem(1)
+        |> Map.put("kid", "test-kid")
+
+      keyset_url = "https://some-tool.example.edu/.well-known/jwks.json"
+
+      insert(:platform_instance, %{
+        client_id: client_id,
+        keyset_url: keyset_url
+      })
+
+      MockHTTP
+      |> expect(:get, fn ^keyset_url ->
+        {:ok,
+         %HTTPoison.Response{
+           status_code: 200,
+           body: Jason.encode!(%{"keys" => [public_jwk]})
+         }}
+      end)
+
+      # Generate JWT with no content items
+      jwt =
+        generate_deep_linking_jwt(client_id, key_pem, "test-kid", %{
+          "https://purl.imsglobal.org/spec/lti-dl/claim/content_items" => []
+        })
+
+      params = %{
+        "JWT" => jwt
+      }
+
+      conn = post(conn, ~p"/lti/deep_link/#{section.slug}/#{activity_id}", params)
+      resp = html_response(conn, 400)
+
+      assert_deep_link_error_response(resp, "Expected exactly one content item, got 0")
+    end
+
+    test "returns 400 for content item with wrong type", %{
+      conn: conn,
+      section: section,
+      activity_id: activity_id
+    } do
+      client_id = "test-client-id"
+      key_pem = File.read!("test/support/fixtures/test_rsa_private.pem")
+
+      public_jwk =
+        JOSE.JWK.from_pem(key_pem)
+        |> JOSE.JWK.to_public()
+        |> JOSE.JWK.to_map()
+        |> elem(1)
+        |> Map.put("kid", "test-kid")
+
+      keyset_url = "https://some-tool.example.edu/.well-known/jwks.json"
+
+      insert(:platform_instance, %{
+        client_id: client_id,
+        keyset_url: keyset_url
+      })
+
+      MockHTTP
+      |> expect(:get, fn ^keyset_url ->
+        {:ok,
+         %HTTPoison.Response{
+           status_code: 200,
+           body: Jason.encode!(%{"keys" => [public_jwk]})
+         }}
+      end)
+
+      # Generate JWT with wrong content item type
+      jwt =
+        generate_deep_linking_jwt(client_id, key_pem, "test-kid", %{
+          "https://purl.imsglobal.org/spec/lti-dl/claim/content_items" => [
+            %{
+              "type" => "file",
+              "title" => "Test File",
+              "url" => "https://example.com/file.pdf"
+            }
+          ]
+        })
+
+      params = %{
+        "JWT" => jwt
+      }
+
+      conn = post(conn, ~p"/lti/deep_link/#{section.slug}/#{activity_id}", params)
+      resp = html_response(conn, 400)
+
+      assert_deep_link_error_response(
+        resp,
+        "Expected content item type to be &#39;ltiResourceLink&#39"
+      )
+    end
+
+    test "returns 400 for non-existent section", %{conn: conn, activity_id: activity_id} do
+      client_id = "test-client-id"
+      key_pem = File.read!("test/support/fixtures/test_rsa_private.pem")
+
+      public_jwk =
+        JOSE.JWK.from_pem(key_pem)
+        |> JOSE.JWK.to_public()
+        |> JOSE.JWK.to_map()
+        |> elem(1)
+        |> Map.put("kid", "test-kid")
+
+      keyset_url = "https://some-tool.example.edu/.well-known/jwks.json"
+
+      insert(:platform_instance, %{
+        client_id: client_id,
+        keyset_url: keyset_url
+      })
+
+      MockHTTP
+      |> expect(:get, fn ^keyset_url ->
+        {:ok,
+         %HTTPoison.Response{
+           status_code: 200,
+           body: Jason.encode!(%{"keys" => [public_jwk]})
+         }}
+      end)
+
+      jwt = generate_deep_linking_jwt(client_id, key_pem, "test-kid")
+
+      params = %{
+        "JWT" => jwt
+      }
+
+      conn = post(conn, ~p"/lti/deep_link/non-existent-section/#{activity_id}", params)
+      resp = html_response(conn, 400)
+
+      assert_deep_link_error_response(resp, "section_not_found")
+    end
+
+    @tag capture_log: true
+    test "returns 400 for JWT without platform instance", %{
+      conn: conn,
+      section: section,
+      activity_id: activity_id
+    } do
+      client_id = "non-existent-client-id"
+      key_pem = File.read!("test/support/fixtures/test_rsa_private.pem")
+
+      jwt = generate_deep_linking_jwt(client_id, key_pem, "test-kid")
+
+      params = %{
+        "JWT" => jwt
+      }
+
+      conn = post(conn, ~p"/lti/deep_link/#{section.slug}/#{activity_id}", params)
+      resp = html_response(conn, 400)
+
+      assert_deep_link_error_response(resp, "invalid_deep_linking_jwt")
     end
   end
 
@@ -624,5 +968,113 @@ defmodule OliWeb.LtiControllerTest do
            })
        }}
     end
+  end
+
+  defp create_lti_external_tool_activity() do
+    attrs = %{
+      "client_id" => "some client_id",
+      "custom_params" => "some custom_params",
+      "description" => "some description",
+      "keyset_url" => "some keyset_url",
+      "login_url" => "some login_url",
+      "name" => "some name",
+      "redirect_uris" => "some redirect_uris",
+      "target_link_uri" => "some target_link_uri"
+    }
+
+    PlatformExternalTools.register_lti_external_tool_activity(attrs)
+  end
+
+  def setup_section(%{conn: conn}) do
+    {:ok, seeds} = setup_project(%{conn: conn})
+
+    {:ok, pub1} = Publishing.publish_project(seeds.project, "some changes", seeds.author.id)
+
+    {:ok, section} =
+      Sections.create_section(%{
+        title: "3",
+        registration_open: true,
+        open_and_free: true,
+        context_id: UUID.uuid4(),
+        institution_id: seeds.institution.id,
+        base_project_id: seeds.project.id,
+        analytics_version: :v1
+      })
+      |> then(fn {:ok, section} -> section end)
+      |> Sections.create_section_resources(pub1)
+
+    student = user_fixture(%{independent_learner: false})
+
+    Sections.enroll(student.id, section.id, [
+      Lti_1p3.Roles.ContextRoles.get_role(:context_learner)
+    ])
+
+    conn =
+      log_in_user(
+        conn,
+        student
+      )
+
+    {:ok,
+     Map.merge(seeds, %{
+       conn: conn,
+       section: section,
+       student: student
+     })}
+  end
+
+  def setup_project(%{conn: conn}) do
+    {:ok, {_platform_instance, activity_registration, _deployment}} =
+      create_lti_external_tool_activity()
+
+    seeds = Oli.Seeder.base_project_with_resource2()
+
+    project = Map.get(seeds, :project)
+    revision = Map.get(seeds, :revision1)
+    author = Map.get(seeds, :author)
+
+    content = %{
+      "openInNewTab" => "true",
+      "authoring" => %{
+        "parts" => []
+      }
+    }
+
+    {:ok, {%{slug: slug, resource_id: activity_id}, _}} =
+      ActivityEditor.create(project.slug, activity_registration.slug, author, content, [])
+
+    {:ok, {%{slug: slug2, resource_id: activity_id2}, _}} =
+      ActivityEditor.create(project.slug, activity_registration.slug, author, content, [])
+
+    seeds = Map.put(seeds, :activity_id, activity_id) |> Map.put(:activity_id2, activity_id2)
+
+    update = %{
+      "content" => %{
+        "version" => "0.1.0",
+        "model" => [
+          %{
+            "type" => "activity-reference",
+            "id" => "1",
+            "activitySlug" => slug
+          },
+          %{
+            "type" => "activity-reference",
+            "id" => "2",
+            "activitySlug" => slug2
+          }
+        ]
+      }
+    }
+
+    PageEditor.acquire_lock(project.slug, revision.slug, author.email)
+    assert {:ok, _} = PageEditor.edit(project.slug, revision.slug, author.email, update)
+
+    conn =
+      log_in_author(
+        conn,
+        seeds.author
+      )
+
+    {:ok, Map.merge(%{conn: conn}, seeds)}
   end
 end
